pyproject.toml
[tool.poetry]
name = "vcfscraper"
version = "0.1.0"
description = "Sample code for scraping with playwright"
authors = ["Vincent Codes Finance <vincent@codes.finance>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
pytest-playwright = "^0.5.0"
requests = "^2.32.3"
pandas = "^2.2.2"
lxml = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
